refactor(query): give orderability a name a test can reach
Orderability, the relation ORDER BY and the aggregations read, sat in an
anonymous namespace inside a header shared by nine translation units. Internal
linkage meant nothing outside those translation units could name it, so the
relation could only be exercised through a plan: the one test covering it
stands up a storage accessor, a symbol table and three operators in order to
call a comparison function. It could not be benchmarked or profiled on its own
at all.

It moves to query/value_relations.hpp under the name openCypher gives it,
alongside a note of what the other three relations are and why the four are
only comprehensible as a set. The definitions stay in the header because a sort
calls them once per comparison, which is the constraint that put them there in
the first place and the one a later tidy-up is most likely to undo. Behaviour
is unchanged; the callers now say which relation they are reading.
